+++ MOBAC/src/main/java/mobac/program/model/Atlas.java
@@ -104,8 +104,8 @@ public class Atlas implements AtlasInterface, ToolTipProvider, TreeNode {
 		return layers.iterator();
 	}
 
-	public int calculateTilesToDownload() {
-		int tiles = 0;
+	public long calculateTilesToDownload() {
+		long tiles = 0;
 		for (LayerInterface layer : layers)
 			tiles += layer.calculateTilesToDownload();
 		return tiles;
